+++ tools/xdhwebq/xdwrgstry.cpp
@@ -22,7 +22,3 @@
 using namespace xdwrgstry;
 
 entry___ xdwrgstry::parameter::ModuleFilename( "ModuleFilename", sclrgstry::Parameters );
-
-
-
-